Safari-609.1.11.1/LayoutTests/js/dom/Promise-resolve-state-in-workers-expected.txt

[Worker] Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/Promise-resolve-state-in-workers.js
PASS [Worker] promiseState is "pending"
PASS [Worker] promiseState is "pending"
PASS [Worker] promiseState is "fulfilled"
PASS [Worker] promiseResult is "hello"
PASS successfullyParsed is true

TEST COMPLETE
